requirements.txt
fastapi
uvicorn
pymongo
beautifulsoup4
requests
textblob
python-dotenv
selenium